Only provide LED strip recess on inside of ears
Also leave a gap between the headband and the ear to thread wires/LED
strip through. This is still a bit narrower than it should be (risk of
breaking soldered connections on LED strip), but already works.

This makes the attachment of the ears to the headband significantly
weaker, and should be compensated for by filling the gap (or the
entire ear for diffusion purposes) with glue once the electronics are
installed.

This is the version used for the cat ears shown in [1].

[1]: https://chaos.social/@TheUnicornConsortium/113367259535493442